fix skill list related population bug, change population counts from int to long, improve nav bar ui, reduce spinner size

+++ peerpedia-server/src/main/java/com/jomin/peerpedia/controller/AdminController.java
@@ -45,7 +45,7 @@ public class AdminController {
             return new ResponseEntity<>(responseBody, HttpStatus.BAD_REQUEST);
         }
 
-        Map<String, Integer> populationResponse = populationService.populate();
+        Map<String, Long> populationResponse = populationService.populate();
         statsService.updateStats();
 
         responseBody.put("success", true);
